requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1.7
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.5
isodate==0.6.0
itypes==1.2.0
lxml==4.6.2
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
zeep==4.0.0
